PWGUD: Add possibility to produce UPC candidates using only ITS-TPC tracks (#1595)
* Add possibility to produce UPC candidates using only ITS-TPC tracks

* Fix megalinter warnings for relevant files

* clang-format
