+++ PWGUD/Tasks/CMakeLists.txt
@@ -43,3 +43,8 @@ o2physics_add_dpl_workflow(upccand-analyzer
                            SOURCES UPCCandidateAnalyzer.cxx
                            PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2::DetectorsBase
                            COMPONENT_NAME Analysis)
+
+o2physics_add_dpl_workflow(upccand-producer-qa
+                           SOURCES UPCCandidateProducerQa.cpp
+                           PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2::DetectorsBase
+                           COMPONENT_NAME Analysis)
